Moving liberty parser to crlcore
