+++ crlcore/meson.build
@@ -2,6 +2,7 @@ subdir('src/ccore')
 subdir('src/LibraryManager')
 subdir('src/cyclop')
 subdir('src/pyCRL')
+subdir('src/liberty')
 subdir('python')
 subdir('doc/crlcore')
