+++ crlcore/src/liberty/Statement.cpp
@@ -14,11 +14,11 @@
 // +-----------------------------------------------------------------+
 
 #include <iostream>
-#include "hurricane/liberty/Attribute.h"
-#include "hurricane/liberty/Define.h"
-#include "hurricane/liberty/Group.h"
-#include "hurricane/liberty/Statement.h"
-#include "hurricane/liberty/Library.h"
+#include "crlcore/liberty/Attribute.h"
+#include "crlcore/liberty/Define.h"
+#include "crlcore/liberty/Group.h"
+#include "crlcore/liberty/Statement.h"
+#include "crlcore/liberty/Library.h"
 
 namespace Liberty {
