+++ hurricane/src/meson.build
@@ -4,8 +4,7 @@ hurricane_includes = include_directories(
   'isobar',
   'configuration',
   'analog',
-  'viewer',
-  'liberty'
+  'viewer'
 )
 
 hurricane_deps = [qt_deps, py_deps, rapidjson, boost]
@@ -16,7 +15,6 @@ subdir('isobar')
 subdir('configuration')
 subdir('analog')
 subdir('viewer')
-subdir('liberty')
 
 py.extension_module(
   'Hurricane',
@@ -29,7 +27,7 @@ py.extension_module(
 )
 
 Hurricane = declare_dependency(
-  link_with: [viewer,analog,configuration,isobar,utilities,hurricane,liberty],
+  link_with: [viewer,analog,configuration,isobar,utilities,hurricane],
   include_directories: hurricane_includes,
   dependencies: [qt_deps, py_deps, boost, rapidjson]
 )
